% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/SA23204184.R
\name{fun}
\alias{fun}
\title{function sin(x)}
\usage{
fun(x)

fun(x)
}
\arguments{
\item{x}{the value of the independent variable}
}
\value{
the value of sin(x)

the value of sin(x)
}
\description{
function sin(x)

function sin(x)
}
\examples{
\dontrun{
sin(2)
}
\dontrun{
sin(2)
}
}
