% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/SA23204184.R
\name{LeftRect}
\alias{LeftRect}
\title{calculate the integral of the function}
\usage{
LeftRect(down, up, n)

LeftRect(down, up, n)
}
\arguments{
\item{down}{the value of the lower bound}

\item{up}{the value of the higher bound}

\item{n}{the number of intervals divided}
}
\value{
the integral

the integral
}
\description{
calculate the integral of the function

calculate the integral of the function
}
\examples{
\dontrun{
    LeftRect(1,2,100)
}
\dontrun{
    LeftRect(1,2,100)
}
}
